requirements.txt
wheel
requests==2.28.1
allure-behave==2.9.45
allure-python-commons==2.9.45
allure-robotframework==2.9.45
Appium-Python-Client==2.2.0
behave==1.2.6
chromedriver-autoinstaller==0.3.1
geckodriver-autoinstaller==0.1.0
edgedriver-autoinstaller==0.2.1
robotframework==5.0.1
selenium==4.1.5
